@@ -10,7 +10,7 @@ authors = [
 ]
 description = "Binary data structures (un-)packing library"
 readme = {file = "README.rst", content-type = "text/x-rst"}
-requires-python = ">=3.9"
+requires-python = ">=3.10"
 keywords = [
     "binary",
     "struct",
