@@ -1,6 +1,5 @@
 """Test bpack (record) descriptors."""
 
-import typing
 import dataclasses
 from collections.abc import Sequence
 
@@ -227,7 +226,7 @@ def test_inconsistent_field_type_and_signed():
 def test_repeat_typing():
     @bpack.descriptor
     class Record:
-        field_1: typing.List[int] = bpack.field(size=4, default=0, repeat=2)
+        field_1: list[int] = bpack.field(size=4, default=0, repeat=2)
         field_2: float = bpack.field(size=8, default=1 / 3)
 
     assert bpack.calcsize(Record()) == 16
@@ -247,7 +246,7 @@ def test_no_repeat_typing():
 
         @bpack.descriptor
         class Record:
-            field_1: typing.List[int] = bpack.field(size=4, default=0)
+            field_1: list[int] = bpack.field(size=4, default=0)
             field_2: float = bpack.field(size=8, default=1 / 3)
 
 
@@ -306,9 +305,7 @@ class TestExplicitSize:
 
             @bpack.descriptor(size=12)
             class Record:
-                field_1: typing.List[int] = bpack.field(
-                    size=4, default=0, repeat=2
-                )
+                field_1: list[int] = bpack.field(size=4, default=0, repeat=2)
                 field_2: float = bpack.field(size=8, default=1 / 3)
 
     @staticmethod
@@ -327,7 +324,7 @@ class TestExplicitSize:
             @bpack.descriptor(size=20)
             class Record:
                 field_1: int = bpack.field(size=4, default=0)
-                field_2: typing.List[float] = bpack.field(
+                field_2: list[float] = bpack.field(
                     size=8, default=1 / 3, offset=8, repeat=2
                 )
 
@@ -398,9 +395,7 @@ class TestSize:
     def test_len_with_repeat_01_typing():
         @bpack.descriptor
         class Record:
-            field_1: typing.List[int] = bpack.field(
-                size=4, default=0, repeat=2
-            )
+            field_1: list[int] = bpack.field(size=4, default=0, repeat=2)
             field_2: float = bpack.field(size=8, default=1 / 3)
 
         assert bpack.calcsize(Record) == 16
@@ -418,12 +413,8 @@ class TestSize:
     def test_len_with_repeat_02_typing():
         @bpack.descriptor
         class Record:
-            field_1: typing.List[int] = bpack.field(
-                size=4, default=0, repeat=2
-            )
-            field_2: typing.List[float] = bpack.field(
-                size=8, default=1 / 3, repeat=2
-            )
+            field_1: list[int] = bpack.field(size=4, default=0, repeat=2)
+            field_2: list[float] = bpack.field(size=8, default=1 / 3, repeat=2)
 
         assert bpack.calcsize(Record) == 24
 
@@ -440,12 +431,10 @@ class TestSize:
     def test_len_with_repeat_and_offset_01_typing():
         @bpack.descriptor
         class Record:
-            field_1: typing.List[int] = bpack.field(
+            field_1: list[int] = bpack.field(
                 size=4, default=0, repeat=2, offset=6
             )
-            field_2: typing.List[float] = bpack.field(
-                size=8, default=1 / 3, repeat=2
-            )
+            field_2: list[float] = bpack.field(size=8, default=1 / 3, repeat=2)
 
         assert bpack.calcsize(Record) == 30
 
@@ -464,10 +453,8 @@ class TestSize:
     def test_len_with_repeat_and_offset_02_typing():
         @bpack.descriptor
         class Record:
-            field_1: typing.List[int] = bpack.field(
-                size=4, default=0, repeat=2
-            )
-            field_2: typing.List[float] = bpack.field(
+            field_1: list[int] = bpack.field(size=4, default=0, repeat=2)
+            field_2: list[float] = bpack.field(
                 size=8, default=1 / 3, repeat=2, offset=14
             )
 
@@ -490,10 +477,8 @@ class TestSize:
 
             @bpack.descriptor
             class Record:
-                field_1: typing.List[int] = bpack.field(
-                    size=4, default=0, repeat=2
-                )
-                field_2: typing.List[float] = bpack.field(
+                field_1: list[int] = bpack.field(size=4, default=0, repeat=2)
+                field_2: list[float] = bpack.field(
                     size=8, default=1 / 3, repeat=2, offset=6
                 )
 
@@ -520,10 +505,10 @@ def test_sequence_type():
 
     @bpack.descriptor
     class Record:
-        field_1: typing.List[int] = bpack.field(size=1, repeat=4)
+        field_1: list[int] = bpack.field(size=1, repeat=4)
 
     field_1 = bpack.fields(Record)[0]
-    assert field_1.type is typing.List[int]
+    assert field_1.type == list[int]
     assert bpack.calcsize(Record) == 4
 
     @bpack.descriptor
@@ -543,7 +528,7 @@ def test_sequence_type():
 
         @bpack.descriptor
         class Record:
-            field_1: typing.Tuple[int, int] = bpack.field(size=1, repeat=2)
+            field_1: tuple[int, int] = bpack.field(size=1, repeat=2)
 
     with pytest.raises(TypeError):
 
