@@ -76,9 +76,9 @@ def test_effective_type():
         type[typing.Any],
         tuple[int, float],
         tuple[int],
-        typing.Type[typing.Any],
-        typing.Tuple[int, float],
-        typing.Tuple[int],
+        type[typing.Any],
+        tuple[int, float],
+        tuple[int],
     ):
         assert bpack.utils.effective_type(type_) == type_
 
@@ -115,11 +115,11 @@ def test_effective_type_from_seq():
     assert bpack.utils.effective_type(list[str]) is str
     assert bpack.utils.effective_type(list[bytes]) is bytes
     assert bpack.utils.effective_type(list[float]) is float
-    assert bpack.utils.effective_type(typing.List[bool]) is bool
-    assert bpack.utils.effective_type(typing.List[int]) is int
-    assert bpack.utils.effective_type(typing.List[str]) is str
-    assert bpack.utils.effective_type(typing.List[bytes]) is bytes
-    assert bpack.utils.effective_type(typing.List[float]) is float
+    assert bpack.utils.effective_type(list[bool]) is bool
+    assert bpack.utils.effective_type(list[int]) is int
+    assert bpack.utils.effective_type(list[str]) is str
+    assert bpack.utils.effective_type(list[bytes]) is bytes
+    assert bpack.utils.effective_type(list[float]) is float
 
 
 def test_effective_type_from_annotated_type():
@@ -145,7 +145,7 @@ def test_get_sequence_type():
     type_ = list[int]
     assert bpack.utils.sequence_type(type_) is list
 
-    type_ = typing.List[int]
+    type_ = list[int]
     assert bpack.utils.sequence_type(type_) is list
 
     type_ = typing.Sequence[int]
@@ -156,18 +156,18 @@ def test_get_sequence_type():
     with pytest.raises(TypeError):
         bpack.utils.sequence_type(type_, error=True)
 
-    type_ = typing.Tuple[int, int]
+    type_ = tuple[int, int]
     assert bpack.utils.sequence_type(type_) is None
     with pytest.raises(TypeError):
         bpack.utils.sequence_type(type_, error=True)
 
     assert bpack.utils.sequence_type(list) is None
-    assert bpack.utils.sequence_type(typing.List) is None
+    assert bpack.utils.sequence_type(list) is None
     assert bpack.utils.sequence_type(typing.Sequence) is None
     assert bpack.utils.sequence_type(tuple) is None
-    assert bpack.utils.sequence_type(typing.Tuple) is None
+    assert bpack.utils.sequence_type(tuple) is None
     assert bpack.utils.sequence_type(type[typing.Any]) is None
-    assert bpack.utils.sequence_type(typing.Type[typing.Any]) is None
+    assert bpack.utils.sequence_type(type[typing.Any]) is None
 
 
 def test_get_sequence_type_from_annotated_type():
@@ -176,7 +176,7 @@ def test_get_sequence_type_from_annotated_type():
     type_ = list[bpack.T[typestr]]
     assert bpack.utils.sequence_type(type_) is list
 
-    type_ = typing.List[bpack.T[typestr]]
+    type_ = list[bpack.T[typestr]]
     assert bpack.utils.sequence_type(type_) is list
 
     type_ = typing.Sequence[bpack.T[typestr]]
@@ -187,7 +187,7 @@ def test_is_sequence_type():
     type_ = list[int]
     assert bpack.utils.is_sequence_type(type_)
 
-    type_ = typing.List[int]
+    type_ = list[int]
     assert bpack.utils.is_sequence_type(type_)
 
     type_ = typing.Sequence[int]
@@ -198,16 +198,16 @@ def test_is_sequence_type():
     with pytest.raises(TypeError):
         bpack.utils.is_sequence_type(type_, error=True)
 
-    type_ = typing.Tuple[int, int]
+    type_ = tuple[int, int]
     assert not bpack.utils.is_sequence_type(type_)
     with pytest.raises(TypeError):
         bpack.utils.is_sequence_type(type_, error=True)
 
     assert not bpack.utils.is_sequence_type(list)
-    assert not bpack.utils.is_sequence_type(typing.List)
+    assert not bpack.utils.is_sequence_type(list)
     assert not bpack.utils.is_sequence_type(typing.Sequence)
     assert not bpack.utils.is_sequence_type(tuple)
-    assert not bpack.utils.is_sequence_type(typing.Tuple)
+    assert not bpack.utils.is_sequence_type(tuple)
 
 
 def test_is_sequence_type_from_annotated_type():
@@ -216,7 +216,7 @@ def test_is_sequence_type_from_annotated_type():
     type_ = list[bpack.T[typestr]]
     assert bpack.utils.is_sequence_type(type_)
 
-    type_ = typing.List[bpack.T[typestr]]
+    type_ = list[bpack.T[typestr]]
     assert bpack.utils.is_sequence_type(type_)
 
     type_ = typing.Sequence[bpack.T[typestr]]
